backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore==2.16.0
google-generativeai>=0.8.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
